% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ksarakil}
\alias{ksarakil}
\title{Ksâr 'Akil Radiocarbon Dates}
\format{
A \code{\link{data.frame}} with 16 rows and 3 variables:
\describe{
\item{Name}{Laboratory code.}
\item{Date}{Radiocarbon date (year BP).}
\item{Uncertainty}{Radiocarbon error (year).}
}
}
\source{
Bosch, M. D., Mannino, M. A., Prendergast, A. L., O'Connell, T. C.,
Demarchi, B., Taylor, S. M., Niven, L., van der Plicht, J. and Hublin, J.-J.
(2015). New Chronology for Ksâr 'Akil (Lebanon) Supports Levantine Route of
Modern Human Dispersal into Europe. \emph{Proceedings of the National Academy of
Sciences} 112(25): 7683-8. \doi{10.1073/pnas.1501529112}.
}
\usage{
ksarakil
}
\description{
Ksâr 'Akil Radiocarbon Dates
}
\keyword{datasets}
